lisp: address ps2/input-service review; document the async server contract
Code-reviewer fixes on the ps2 migration (slice 1b):
- %ps2-poll now returns an encoded fixnum (scancode*2 + pressed) and the Lisp
  pump decodes it, instead of building a list in C. The primitive allocates
  nothing, eliminating the GC-rooting hazard where a half-built list held only
  in a C local could be swept by an inline OOM collect of the per-context heap.
- g_irq_hook is now volatile (read from the ISR; the compiler could otherwise
  cache the NULL it sees post-init).
- The keyboard IRQ fires the wake hook only when a real key event was queued,
  not on every mouse interrupt (no spurious wake/re-park churn).
- PS2_Initialize returns int (the failure sentinel is -1; uint8_t truncated it),
  now that module_init gates on the return value.
- Removed the redundant interrupt_setmask() calls (interrupt_mapinterrupt
  already clears the RTE mask bit) and deleted the dead #if 0 scancode-set-3
  table (it referenced types removed with the CoreInput include).
- Documented why g_ps2_waiter can't dangle (the context object lives in the
  frozen system heap and finished contexts aren't reaped) and why publishing it
  before cli() is race-free.
- run-qemu.sh SENDKEY: auto-pick a free VNC display (scan upward) and clean up
  the monitor socket on exit.

Also adds notes/servers/lisp-server-migration.md: the async actor contract
(register/event, request/complete), the ISR->wake->poll->send driver loop, the
current constraints (BSP-pinned, small-messages-only until a bytevector message
type lands, decide coordinate-vs-rewrite per heavy server), and the migration
order -- the template for moving CoreStorage/CoreNetwork/etc. to Lisp.

Verified: host suite green; QEMU keystrokes still decode to (key N 1)/(key N 0)
events on the coreinput context; clean boot; no faults.

+++ drivers/ps2/inc/ps2.h
@@ -14,7 +14,7 @@
 #define BUF_LEN 8192
 #define ENT_SIZE 4
 
-uint8_t PS2_Initialize();
+int PS2_Initialize();
 uint8_t PS2_ReadStatus();
 uint8_t PS2_ReadConfig();
 void PS2_WriteConfig(uint8_t cfg);
